0.9.0 beta e
Updated APIs and mod.

+++ src/main/java/buildcraft/api/power/ILaserTarget.java
@@ -1,18 +1,13 @@
-/*
- * Copyright (c) SpaceToad, 2011-2012
+/**
+ * Copyright (c) 2011-2014, SpaceToad and the BuildCraft Team
  * http://www.mod-buildcraft.com
- * 
- * BuildCraft is distributed under the terms of the Minecraft Mod Public
- * License 1.0, or MMPL. Please check the inventory of the license located in
- * http://www.mod-buildcraft.com/MMPL-1.0.txt
+ *
+ * The BuildCraft API is distributed under the terms of the MIT License.
+ * Please check the contents of the license, which should be located
+ * as "LICENSE.API" in the BuildCraft source code distribution.
  */
 package buildcraft.api.power;
 
-/**
- * Specifies a Tile Entity that can receive power via laser beam.
- *
- * @author cpw
- */
 public interface ILaserTarget {
 
 	/**
@@ -28,7 +23,7 @@ public interface ILaserTarget {
 	 *
 	 * @param energy
 	 */
-	void receiveLaserEnergy(float energy);
+	void receiveLaserEnergy(int energy);
 
 	/**
 	 * Return true if the Tile Entity object is no longer a valid target. For
@@ -38,9 +33,9 @@ public interface ILaserTarget {
 	 */
 	boolean isInvalidTarget();
 
-	int getXCoord();
+	double getXCoord();
 
-	int getYCoord();
+	double getYCoord();
 
-	int getZCoord();
+	double getZCoord();
 }
